+++ CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
+### Added
+- enable griffon service products-contains-component --search-community
+- enable griffon entities community-component-registry
 
 ## [0.1.16] - 2023-04-11
 ### Changed